Azure Projects/Azure learn/requirements.txt
azure-functions
requests